[Homework] добавление БД. + CashProvider

+++ app/src/main/java/ru/androidschool/intensiv/data/actors/ActorsRepository.kt
@@ -4,5 +4,5 @@ import io.reactivex.Single
 import ru.androidschool.intensiv.network.entity.CreditsResponse
 
 interface ActorsRepository {
-    fun getActors(id: Int): Single<CreditsResponse>
+    fun getActors(id: Long): Single<CreditsResponse>
 }
