+++ app/src/main/java/ru/androidschool/intensiv/ui/details/MovieDetailsViewModelFactory.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
 @Suppress("UNCHECKED_CAST")
-class MovieDetailsViewModelFactory(private val movieId: Int?) :
+class MovieDetailsViewModelFactory(private val movieId: Long?) :
     ViewModelProvider.NewInstanceFactory() {
 
     override fun <T : ViewModel?> create(modelClass: Class<T>): T =
